Refac/Fix: Create random generator at *function* call. Read description.
The reason why this change is made is because randomness is a prevalent aspect of the social systems at hand. Therefore, even if any "social system" object gets copied (e.g. with `deepcopy`), then any subsequent computations done with them should result to different outputs (due to the random generators having a different seed).
On the other hand, if the random generator object was (deep) copied along with everything else in the object in the current implementation, then both instances of the "social system" objects would produce the same (pseudo-)random results, since the random generators would be identical (including their seed).