Refac: Split functionalities and definitions of `main.py` to other files
A. DEFINITIONS
=============
1. Experimentation constants `nr_experiments`, `nr_agents` and `time_steps` at `src/experiments/__init__.py`

2. Social system (constant) parameters in `src/social_system/utils/help_constant_structures.py`. So far, this only contains the different μ and θ configurations that will be used.

B. FUNCTIONALITIES
=================
1. `prepare_configurations.py`.
The different experiment configurations that will be conducted for the project will be created here.
It is incomplete, as constants for different `expected_edges_per_agent`, as well as `(pr_friend, pr_friend_of_friend)` have not been defined (yet). These would be defined in the `help_constant_structures.py` (probably?).