Feat + Doc: New dataclass for `run_comparison_experiments`, doc for fcn.
A somewhat new decision is the fact that metric of the system will not be part of the `pandas.DataFrame`. This is because (currently) the implemented metrics are dependent on the opinions _alone_. Therefore, metrics can be computed by an external function call (implementing the same metric).

This will need refactoring if metrics requiring more information are devised. A potential solution (though not well-thought through) is that the `run_experiment` function could additionally return the final state of the `BackboneSocialSystem` and `HomophilicSocialSystem`, similarly to how the initial configurations of said systems are returned in the `run_comparison_experiments` function.